[PC/SC test] Test repeated card insert/remove
Add a test for repeating cycles of card insertion and removal.

Besides checking one of the basic scenarios of SCardGetStatusChange(),
this also verifies that quickly succeeding reader init/deinit events
don't lead to internal breakages. Specifically, this test has been
used to find a racy use-after-free in the Libusb subsystem: #1098.